sistema_experto/requirements.txt
clipspy>=1.0.6
flask>=3.0.0
flask-cors>=4.0.0
